requirements.txt
python-dotenv
py-cord
replicate